+++ ent/proxyredemptionrecord/where.go
@@ -76,7 +76,7 @@ func UnitTime(v string) predicate.ProxyRedemptionRecord {
 }
 
 // Quantity applies equality check predicate on the "quantity" field. It's identical to QuantityEQ.
-func Quantity(v int64) predicate.ProxyRedemptionRecord {
+func Quantity(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldEQ(FieldQuantity, v))
 }
 
@@ -236,42 +236,42 @@ func UnitTimeContainsFold(v string) predicate.ProxyRedemptionRecord {
 }
 
 // QuantityEQ applies the EQ predicate on the "quantity" field.
-func QuantityEQ(v int64) predicate.ProxyRedemptionRecord {
+func QuantityEQ(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldEQ(FieldQuantity, v))
 }
 
 // QuantityNEQ applies the NEQ predicate on the "quantity" field.
-func QuantityNEQ(v int64) predicate.ProxyRedemptionRecord {
+func QuantityNEQ(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldNEQ(FieldQuantity, v))
 }
 
 // QuantityIn applies the In predicate on the "quantity" field.
-func QuantityIn(vs ...int64) predicate.ProxyRedemptionRecord {
+func QuantityIn(vs ...int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldIn(FieldQuantity, vs...))
 }
 
 // QuantityNotIn applies the NotIn predicate on the "quantity" field.
-func QuantityNotIn(vs ...int64) predicate.ProxyRedemptionRecord {
+func QuantityNotIn(vs ...int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldNotIn(FieldQuantity, vs...))
 }
 
 // QuantityGT applies the GT predicate on the "quantity" field.
-func QuantityGT(v int64) predicate.ProxyRedemptionRecord {
+func QuantityGT(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldGT(FieldQuantity, v))
 }
 
 // QuantityGTE applies the GTE predicate on the "quantity" field.
-func QuantityGTE(v int64) predicate.ProxyRedemptionRecord {
+func QuantityGTE(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldGTE(FieldQuantity, v))
 }
 
 // QuantityLT applies the LT predicate on the "quantity" field.
-func QuantityLT(v int64) predicate.ProxyRedemptionRecord {
+func QuantityLT(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldLT(FieldQuantity, v))
 }
 
 // QuantityLTE applies the LTE predicate on the "quantity" field.
-func QuantityLTE(v int64) predicate.ProxyRedemptionRecord {
+func QuantityLTE(v int32) predicate.ProxyRedemptionRecord {
 	return predicate.ProxyRedemptionRecord(sql.FieldLTE(FieldQuantity, v))
 }
